[Feat] 경기기록권한관리 Modal - 경기기록자 저장 API 구현 완료

+++ src/views/game/recordDetail/modal/GameRecordAuthManageModal.vue
@@ -18,7 +18,10 @@
 			<v-card-title>경기기록 입력권한관리</v-card-title>
 			<v-container>
 				<div class="text-right">
-					<GameRecordAuthSaveBtn pBtnName="저장" @do-save="saveGameRecorders" />
+					<GameRecordAuthSaveBtn
+						pBtnName="저장"
+						@do-save="saveGameRecorders()"
+					/>
 				</div>
 				<v-container>
 					<div>권한자 목록</div>
@@ -49,7 +52,7 @@
 <script>
 	import GameAPI from '@/api/GameAPI.js';
 
-	import { HomeAwayCode } from '@/const/code/GameCode.js';
+	import { HomeAwayCode, GameRecordAuthCode } from '@/const/code/GameCode.js';
 	import ValidationUtil from '@/common/util/ValidationUtil.js';
 	import ArrayUtil from '@/common/util/ArrayUtil.js';
 
@@ -114,8 +117,17 @@
 				this.gameJoinPlayers = res.data.gameJoinTeamMembers;
 				this.isGetGameJoinPlayersLoadOk = true;
 			},
-			saveGameRecorders() {
+			async saveGameRecorders() {
 				// TODO 구현 예정
+				const pathVar = {
+					gameSeq: this.gameSeq,
+				};
+
+				const reqBody = {
+					gameRecorders: this.gameRecorders,
+				};
+
+				await GameAPI.saveGameRecorders(pathVar, reqBody);
 			},
 			addGameRecorder(targetPlayer) {
 				console.log(targetPlayer);
@@ -131,8 +143,8 @@
 					return;
 				}
 				// 경기기록권한 코드값 할당 후 목록에 추가
-				targetPlayer.gameRecordAuthCode = '02';
-				targetPlayer.gameRecordAuthCodeName = '입력권한자';
+				targetPlayer.gameRecordAuthCode = GameRecordAuthCode.ONLY_WRITER;
+				targetPlayer.gameRecordAuthCodeName = '경기기록자';
 				this.gameRecorders.push(targetPlayer);
 			},
 			deleteGameRecorder(targetPlayer) {
@@ -148,6 +160,12 @@
 				if (!hasDeleteTarget) {
 					return;
 				}
+
+				if (GameRecordAuthCode.CREATOR === deleteTarget.gameRecordAuthCode) {
+					alert('경기생성자는 삭제할 수 없습니다.');
+					return;
+				}
+
 				this.gameRecorders = ArrayUtil.deleteItemById(
 					this.gameRecorders,
 					targetPlayer,
